rename: 디렉토리 변경 #12
Main 페이지에서 작업하던 것을 CafeSearchPage로 변경
